Refactor: extract 'IUserType' interface
