+++ src/routes/graphql/api/IUserType.ts
@@ -1,0 +1,9 @@
+export interface IUserType {
+  id: string;
+  name: string;
+  balance: number;
+  profile: unknown;
+  posts: unknown[];
+  userSubscribedTo: IUserType[];
+  subscribedToUser: IUserType[];
+}
